전체 코드

+++ src/components/article/LikeToggleButton.tsx
@@ -1,5 +1,7 @@
 import { useState } from "react";
 import Image from "next/image";
+import { useRouter } from "next/router";
+import { useAuth } from "@context/AuthContext";
 import { deleteLike, postLike } from "@lib/api/articleApi";
 import likeOffIcon from "@icons/ic_heartEmpty.svg";
 import likeOnIcon from "@icons/ic_heartFill.svg";
@@ -17,19 +19,30 @@ const LikeToggleButton = ({
 }: LikeToggleButtonProps) => {
   const [isLiked, setIsLiked] = useState(initialLiked);
   const [likeCount, setLikeCount] = useState(initialLikeCount);
+  const router = useRouter();
+  const { isLoggedIn } = useAuth();
 
   const handleLikeToggle = async () => {
-    try {
-      if (isLiked) {
-        const updatedArticle = await deleteLike(targetId);
-        setLikeCount(updatedArticle.likeCount);
-      } else {
-        const updatedArticle = await postLike(targetId);
-        setLikeCount(updatedArticle.likeCount);
+    if (!isLoggedIn) {
+      const confirmed = window.confirm(
+        "로그인이 필요한 서비스입니다. 로그인 하시겠습니까?",
+      );
+      if (confirmed) {
+        router.push("/login");
       }
-      setIsLiked((prevIsLiked) => !prevIsLiked);
-    } catch (err) {
-      console.error("좋아요 토글 실패", err);
+    } else {
+      try {
+        if (isLiked) {
+          const updatedArticle = await deleteLike(targetId);
+          setLikeCount(updatedArticle.likeCount);
+        } else {
+          const updatedArticle = await postLike(targetId);
+          setLikeCount(updatedArticle.likeCount);
+        }
+        setIsLiked((prevIsLiked) => !prevIsLiked);
+      } catch (err) {
+        console.error("좋아요 토글 실패", err);
+      }
     }
   };
